use a tmpfile instead of an in-memory buffer
The plugin used an in-memory buffer before. This is logically correct,
as we are not really editing a file, but reading only a part of it.
However, becauseb vim is really built to use backing files, stuff like
filetype or encoding recognition did not work. Also, editing multiple
files in chunks was not really easy.

This changes the plugin to use a tempfile with the same extension. That
way updating the content is easyer, and the filetype and encoding
functionality works out of the box. This change simplifies quite a lot
of the buffer-handling logic.

Now also makes the chunk files unmodifiable to prevent accidentally
changing something (this happend to me sometimes).