add support for writing the chunk file and updating the source
Previously, the chunk file, which temporeraly holds part of the contents
of the sourcefile, could not be edited. I however found that I sometimes
need to edit big files as well. This commit adds the ability to modify
the chunk file, and after writing it, the source file will also be
updated with those changes.